Fix #1349, Break up pc-rtems to support generic configuration
