+++ src/bsp/pc-rtems/src/bsp_console.c
@@ -17,9 +17,8 @@
  ************************************************************************/
 
 /*
- * File:  bsp_console.c
+ * \file
  *
- * Purpose:
  *   OSAL BSP debug console abstraction
  */
