+++ src/bsp/pc-rtems/src/pcrtems_bsp_internal.h
@@ -19,7 +19,6 @@
 /**
  * \file
  *
- * Purpose:
  *   Header file for internal data to the PC-RTEMS BSP
  */
